refactor(web): decompose the useChat hook
useChat was one 824-line hook. It keeps the mode and message state, the
chat/conversation selectors, the model-list reconciliation and the
returned surface (488 lines, under the max-lines-per-function cap, so
its eslint-disable goes); four pieces move out next to the hooks already
split from it:

- useAssistantStream.ts: chat mode's single-model streaming
  (streamAssistantReply, handleSend, handleStop, handleRegenerate);
  owns the request abort controller and aborts it on unmount
- useDeleteMessage.ts: handleDeleteMessage with its chat-mode pair rule
  and conversation-mode last-pair rule and state re-derivation
- useChatScroll.ts: the messages container ref, scrollToBottom, the
  scrolled-up intent and the two auto-scroll effects
- chatDerived.ts: lastChatError, failedConversationModel and
  messageTotals as pure functions

useChat() returns the same surface with the same semantics; the 27 Chat
suites pass unchanged. The file leaves the size allowlist (846 -> 488).

Verified on the rebuilt dev stack: the chat page renders, Shift+Enter
inserts a newline without sending, Send stays disabled without a model,
conversation mode's Start stays disabled until both models and a prompt
exist, no console errors.
